Ansh0506, tu toh chud gaya

+++ backend/src/controllers/labController.js
@@ -2,6 +2,8 @@ const { PrismaClient } = require('@prisma/client');
 const { ApiError } = require('../utils/ApiError.js');
 const { ApiResponse } = require('../utils/ApiResponse.js');
 const { asyncHandler } = require('../utils/asyncHandler.js');
+require('dotenv').config()
+const cloudinary = require('cloudinary').v2
 const prisma = new PrismaClient();
 
 const fetchLab = async (req, res) => {
@@ -13,17 +15,17 @@ const fetchLab = async (req, res) => {
 };
 
 const uploadLabMaterial = asyncHandler(async (req, res, next) => {
-  const { title } = req.body;
-  const subjectId = +req.params.sid;
-  if (!req.file) {
-    return next(new ApiError(400, "File is required"));
-    }
-  const url = `/uploads/${req.file.filename}`;
+  const { title, url } = req.body;
+  
+  if(!title || !url) {
+    return next(new ApiError(400, "Title and Cloudinary URL are required"))
+  }
+  
   const labMaterials = await prisma.labMaterial.create({
     data: {
       title,
       url,
-      subjectId,
+      subjectId: +req.params.sid,
       userId: req.user.id
     }
   });
@@ -31,10 +33,21 @@ const uploadLabMaterial = asyncHandler(async (req, res, next) => {
 });
 
 const deleteLabMaterial = asyncHandler(async (req, res, next) => {
-    const labMaterialToDelete = await prisma.labMaterial.delete({
-      where: { id: +req.params.id }
-    });
-    res.status(200).json(new ApiResponse(200, labMaterialToDelete, "lab material deleted successfully"));
+  const file = await prisma.labMaterial.findUnique({
+    where: {id: +req.params.id}
+  })
+
+  if(!file){
+    return res.status(400).json(new ApiError(404, null, "File Not Found"))
+  }
+
+  const public_id = file.title
+  await cloudinary.uploader.destroy(public_id).then(result=>console.log(result))
+
+  const labMaterialToDelete = await prisma.labMaterial.delete({
+    where: { id: +req.params.id }
+  });
+  res.status(200).json(new ApiResponse(200, labMaterialToDelete, "lab material deleted successfully"));
 });
 
 module.exports = {fetchLab , uploadLabMaterial , deleteLabMaterial};
